requirements.txt
email-validator==2.3.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
pytest==9.0.2
